refactor: align public header ownership
