+++ .github/workflows/ci-docker.yml
@@ -24,22 +24,6 @@ env:
       - master
     tags:
       - 'v*'
-    paths:
-      - 'Dockerfile'
-      - '.dockerignore'
-      - 'docker/**'
-      - 'iccdev-mcp/**'
-      - 'Build/**'
-      - 'IccConnect/**'
-      - 'IccJSON/**'
-      - 'IccProfLib/**'
-      - 'IccXML/**'
-      - 'Testing/**'
-      - 'Tools/**'
-      - 'cfl/**'
-      - 'python/**'
-      - '.github/scripts/preflight-safety-checks.sh'
-      - '.github/workflows/ci-docker.yml'
   workflow_dispatch:
 
 concurrency:
@@ -154,7 +138,6 @@ jobs:
             GIT_COMMIT=${{ github.sha }}
           load: ${{ env.PUBLISH_IMAGE != 'true' }}
           push: ${{ env.PUBLISH_IMAGE == 'true' }}
-          cache-to: type=inline
           provenance: ${{ env.PUBLISH_IMAGE == 'true' && 'mode=max' || 'false' }}
           sbom: ${{ env.PUBLISH_IMAGE == 'true' }}
           labels: |
