+++ docs/regression-workflow-governance.md
@@ -154,7 +154,7 @@ workflow update.
 The Docker PR lane consumes the published
 `ghcr.io/internationalcolorconsortium/iccdev:latest` image as a
 maintainer-controlled build cache. It must pull and report the resolved digest,
-then rebuild the checked-out PR Dockerfile with that image as `--cache-from`.
+then rebuild the checked-out PR Dockerfile without a workflow cache.
 The resulting image is local to the job: it must bind the checked-out PR tree
 read-only, copy it to container-local scratch space, and build and run the fast
 CTest envelope there. Only `ci-docker` publishes images.
